requirements.txt
flask
pymongo
flask_pymongo
python-dotenv
json_logging